fix inconsistency between binary and ascii protocols
