+++ src/protocol/mod.rs
@@ -17,6 +17,16 @@ pub enum Protocol {
     Binary(BinaryProtocol),
 }
 
+#[allow(dead_code)]
+pub enum ResponseStatus {
+    NoError = 0x00,
+    KeyNotFound = 0x01,
+    KeyExists = 0x02,
+    ValueTooLarge = 0x03,
+    InvalidArguments = 0x04,
+    AuthenticationRequired = 0x20,
+}
+
 #[enum_dispatch(Protocol)]
 pub trait ProtocolTrait {
     fn auth(&mut self, username: &str, password: &str) -> Result<(), MemcacheError>;
